refactor: move to color
