+++ src/lines.ts
@@ -15,8 +15,9 @@ import {
 import { ICanvasOverlayDrawEvent } from "./canvas-overlay";
 import * as color from "./color";
 import { LineFeatureVertices } from "./line-feature-vertices";
-import { latLngDistance, inBounds, getChosenColor } from "./utils";
+import { latLngDistance, inBounds } from "./utils";
 import glify from "./index";
+import { getChosenColor } from "./color";
 
 export type WeightCallback = (i: number, feature: any) => number;
